+++ packages/ricos-content/web/src/velo-adapter/index.ts
@@ -1,2 +1,3 @@
 export { createBuilder } from './builder';
 export { modify } from './modify';
+export { getText } from './extract';