+++ packages/tiptap-editor/web/src/components/RicosTiptapEditor/RicosTiptapEditor.spec.tsx
@@ -6,6 +6,7 @@ import { compare } from 'ricos-content/lib/comparision';
 import { RichContent } from 'ricos-schema';
 import { pluginDivider } from 'wix-rich-content-plugin-divider';
 import { pluginImage } from 'wix-rich-content-plugin-image';
+import { pluginLink } from 'wix-rich-content-plugin-link';
 import { fromTiptap, toTiptap } from '../../converters';
 import supportedPluginsContent from '../../__tests__/supportedPluginsContent.json';
 import { RicosTiptapEditor } from './RicosTiptapEditor';
@@ -28,7 +29,7 @@ jest.mock('@tiptap/react', () => {
 describe('tiptap editor', () => {
   it('should not change content', async () => {
     const content = toTiptap(RichContent.fromJSON(supportedPluginsContent));
-    const plugins = [pluginImage(), pluginDivider()];
+    const plugins = [pluginImage(), pluginDivider(), pluginLink()];
     const extensions =
       compact(plugins?.flatMap((plugin: TiptapEditorPlugin) => plugin.tiptapExtensions)) || [];
     render(<RicosTiptapEditor content={content} extensions={extensions} t={() => ''} />);